Trait eventually_util::aggregate::referential::Aggregate

pub trait Aggregate: Sized {
    type Event;
    type Error;
    fn apply(self, event: Self::Event) -> Result<Self, Self::Error>;
}

An Aggregate representation where the State is the same as the aggregate root.

Examples

use eventually_util::aggregate::referential::Aggregate;

struct MyState {
    // Some fields
}

enum MyEvent {
    Happened,
}

impl Aggregate for MyState {
    type Event = MyEvent;
    type Error = std::convert::Infallible;

    fn apply(self, event: Self::Event) -> Result<Self, Self::Error> {
        unimplemented!()
    }
}

Associated Types

type Event

Domain events that express Aggregate's mutations.

Check out Event documentation for more information.

type Error

Error returned by apply when handling an incorrect Event.

Check out Error documentation for more information.

Loading content...
Required methods

fn apply(self, event: Self::Event) -> Result<Self, Self::Error>

Applies an Event to the current state of the Aggregate (Self in this instance), returning either the next state or an Error.

Loading content...