Module eventually_util::inmemory

Contains supporting entities using an in-memory backend.

Structs

EventStore
An in-memory EventStore implementation, backed by an HashMap.

EventStoreBuilder
Builder for EventStore instances.

Projector
A Projector manages the state of a single Projection by opening a long-running stream of all events coming from the EventStore.

ProjectorBuilder
Reusable builder for multiple Projector instances.

Enums

Error
Error returned by the EventStore::append when a conflict has been detected.