Trait eventually_util::optional::Aggregate

pub trait Aggregate {
    type Id: Eq;
    type State;
    type Event;
    type Command;
    type Error;
    fn apply_first(event: Self::Event) -> Result<Self::State, Self::Error>;

    fn apply_next(
        state: Self::State, 
        event: Self::Event
    ) -> Result<Self::State, Self::Error>;

    fn handle_first<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        id: &'life1 Self::Id, 
        command: Self::Command
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Self::Event>, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn handle_next<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        id: &'life1 Self::Id, 
        state: &'life2 Self::State, 
        command: Self::Command
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Self::Event>, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    fn into_aggregate(self) -> IntoAggregate<Self>
    where
        Self: Sized,
    { ... }
}

An Option-flavoured, Aggregate-compatible trait to model Aggregates having an optional State.

Use as_aggregate to get an Aggregate-compatible instance of this trait.

Associated Types

type Id: Eq

Identifier type of the Aggregate.

Check out Aggregate::Id for more information.

type State

State of the Aggregate.

Check out Aggregate::State for more information.

type Event

Events produced and supported by the Aggregate.

Check out Aggregate::Event for more information.

type Command

Commands supported by the Aggregate.

Check out Aggregate::Command for more information.

type Error

Error produced by the the Aggregate while applying Events or handling Commands.

Required methods

fn apply_first(event: Self::Event) -> Result<Self::State, Self::Error>

Applies the specified Event when the State is empty.

fn apply_next(
state: Self::State,
event: Self::Event
) -> Result<Self::State, Self::Error>

Applies the specified Event on a pre-existing State value.

fn handle_first<'life0, 'life1, 'async_trait>(
&'life0 self,
id: &'life1 Self::Id,
command: Self::Command
) -> Pin<Box<dyn Future<Output = Result<Vec<Self::Event>, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Handles the specified Commandwhen the State is empty.

fn handle_next<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
id: &'life1 Self::Id,
state: &'life2 Self::State,
command: Self::Command
) -> Pin<Box<dyn Future<Output = Result<Vec<Self::Event>, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Handles the specified Command on a pre-existing State value.

Provided methods

fn into_aggregate(self) -> IntoAggregate<Self> where
Self: Sized,

Translates the current optional::Aggregate instance into a newtype instance compatible with the core Aggregate trait.