Trait eventually_util::optional::CommandHandler

pub trait CommandHandler {
    type Command;
    type Aggregate: Aggregate;
    type Error;
    fn handle_first(
        &self, 
        command: Self::Command
    ) -> BoxFuture<Result<EventOf<Self::Aggregate>, Self::Error>>;

    fn handle_next<'a, 's: 'a>(
        &'a self, 
        state: &'s StateOf<Self::Aggregate>, 
        command: Self::Command
    ) -> BoxFuture<'a, Result<EventOf<Self::Aggregate>, Self::Error>>;

    fn as_handler(self) -> AsHandler<Self>
    where
        Self: Sized,
    { ... }
}

Command Handler trait referring to Aggregate with Option state, a.k.a. Aggregate.

Implementations of this trait can be adapted back into the command::Handler foundation trait by using as_handler, in cases where the implementation has compile-time known size.

Associated Types

type Command

Commands to trigger a specific use-case on the context of an Aggregate.

Most often than not, this type should be an enum containing all supported operations -- that are not queries -- for the specified Aggregate.

type Aggregate: Aggregate

Domain entity produced, updated or, in some way, affected by a Command.

type Error

Possible expected errors to be returned when handling a Command fails.

Loading content...
Required methods

fn handle_first(
&self,
command: Self::Command
) -> BoxFuture<Result<EventOf<Self::Aggregate>, Self::Error>>

Handles a Command when the Aggregate state is not yet present.

Usually this happens when the event store has no persisted event for this aggregate yet.

fn handle_next<'a, 's: 'a>(
&'a self,
state: &'s StateOf<Self::Aggregate>,
command: Self::Command
) -> BoxFuture<'a, Result<EventOf<Self::Aggregate>, Self::Error>>

Handles a Command when the previous Aggregate state is already present and available to the command handler.

Loading content...
Provided methods

fn as_handler(self) -> AsHandler<Self> where
Self: Sized,

Adapts the CommandHandler implementation to the command::Handler foundation trait, useful when needs to be used with a command::Dispatcher.

This method is only available when the Self has compile-time known size.

Loading content...