Trait eventually_util::versioned::AggregateExt

pub trait AggregateExt: Aggregate + Sized {
    fn versioned(self) -> AsAggregate<Self> { ... }
}

Extension trait to add Optimistic Concurrency Control support over an Aggregate using versioning.

Usage

Call versioned over an Aggregate instance to add versioning support.

use eventually_util::versioned::AggregateExt;

// Assuming `SomeAggregateExample` is an Aggregate
let versioned_aggregate = SomeAggregateExample.versioned();

Provided methods

fn versioned(self) -> AsAggregate<Self>

Returns a versioned flavour of an Aggregate.

Loading content...