Module eventually_util::inmemory

Contains supporting entities using an in-memory backend.

Structs

ConflictError
Error returned by the EventStore::append when a conflict has been detected.

EventStore
An in-memory EventStore implementation, backed by an HashMap.

EventStoreBuilder
Builder for EventStore instances.

LaggedError
Error returned by the EventSubscriber when reading elements from the EventStream produced by subscribe_all.

Projector
A Projector manages the state of a single Projection by opening a long-running stream of all events coming from the EventStore.